@@ -2,25 +2,17 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net9.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="StereoKit" Version="0.3.7-preview.4" />
+    <PackageReference Include="StereoKit" Version="0.3.10" />
   </ItemGroup>
 
   <ItemGroup>
     <None Include="Assets\**">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Remove="Assets\floor.hlsl" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <SKShader Include="Assets\floor.hlsl" />
   </ItemGroup>
 
 </Project>
